+++ src/basilisp/lang/compiler/nodes.py
@@ -296,7 +296,6 @@ class LocalType(Enum):
     LET = kw.keyword("let")
     LETFN = kw.keyword("letfn")
     LOOP = kw.keyword("loop")
-    REQUIRE = kw.keyword("require")
     THIS = kw.keyword("this")